Renaming deamontests into daemon/testutils
The cleanest way I thought to ensure the wsl --system mock wouldn't penalyze us on code coverage.
Alternatives are either hackish or involve manually adding exclusion patterns when filtering coverage reports.
I thought it would just be better to name testutils to fill in the exclusion patterns already in use.
